+++ packages/expo-cli/src/commands/publish-info.js
@@ -1,8 +1,9 @@
 /**
  * @flow
  */
+import { readConfigJsonAsync } from '@expo/config';
+import { Api, FormData, Project, UserManager } from '@expo/xdl';
 
-import { Api, FormData, Project, ProjectUtils, UserManager } from '@expo/xdl';
 import * as table from '../commands/utils/cli-table';
 
 const HORIZ_CELL_WIDTH_SMALL = 15;
@@ -36,7 +37,7 @@ export default (program: any) => {
       // TODO(ville): move request from multipart/form-data to JSON once supported by the endpoint.
       let formData = new FormData();
       formData.append('queryType', 'history');
-      let { exp } = await ProjectUtils.readConfigJsonAsync(projectDir);
+      let { exp } = await readConfigJsonAsync(projectDir);
       if (exp.owner) {
         formData.append('owner', exp.owner);
       }
@@ -116,7 +117,7 @@ export default (program: any) => {
       let formData = new FormData();
       formData.append('queryType', 'details');
 
-      let { exp } = await ProjectUtils.readConfigJsonAsync(projectDir);
+      let { exp } = await readConfigJsonAsync(projectDir);
       if (exp.owner) {
         formData.append('owner', exp.owner);
       }