+++ packages/expo-cli/src/commands/publish.js
@@ -1,17 +1,16 @@
 /**
  * @flow
  */
-
+import { readConfigJsonAsync } from '@expo/config';
+import simpleSpinner from '@expo/simple-spinner';
+import { Exp, Project } from '@expo/xdl';
+import chalk from 'chalk';
 import fs from 'fs';
 import path from 'path';
-import chalk from 'chalk';
-import simpleSpinner from '@expo/simple-spinner';
 
-import { Exp, Project, ProjectUtils } from '@expo/xdl';
-
+import { installExitHooks } from '../exit';
 import log from '../log';
 import sendTo from '../sendTo';
-import { installExitHooks } from '../exit';
 
 type Options = {
   clear?: boolean,
@@ -72,7 +71,7 @@ export async function action(projectDir: string, options: Options = {}) {
     sdkVersion,
   });
 
-  const { exp } = await ProjectUtils.readConfigJsonAsync(projectDir);
+  const { exp } = await readConfigJsonAsync(projectDir);
 
   if (
     buildStatus.userHasBuiltExperienceBefore &&